69: Update SDL2 to SDL3 (#70)
* 69: Width and height in PongScene and GeometryTestScene also changed to float for consistency

+++ ai/ai_pong/pong_scene.hpp
@@ -30,12 +30,12 @@ private:
   void initialize(const int width, const int height);
   void finalize();
   void resize(const int width, const int height);
-  void animate(const std::uint32_t time_elapsed_ms);
+  void animate(const std::uint64_t time_elapsed_ms);
   void redraw();
 
-  std::uint32_t last_timestamp_ms_{};
-  int width_{};
-  int height_{};
+  std::uint64_t last_timestamp_ms_{};
+  float width_{};
+  float height_{};
 
   std::vector<GameState> game_states_{number_per_generation};
   std::vector<std::unique_ptr<NeuralNetwork>> neural_networks_{number_per_generation};
