+++ loaders/loaders_assimp/model_scene.hpp
@@ -24,14 +24,14 @@ private:
   void initialize(const int width, const int height);
   void finalize();
   void resize(const int width, const int height);
-  void animate(const std::uint32_t timestamp);
+  void animate(const std::uint64_t timestamp);
   void redraw();
 
   void upload_data();
 
   std::shared_ptr<const Model> model_{};
 
-  std::uint32_t last_timestamp_ms_{};
+  std::uint64_t last_timestamp_ms_{};
 
   glm::mat4 projection_{};
   glm::vec3 camera_pos_{};
