+++ wolf/wolf_common/player_state.hpp
@@ -14,7 +14,7 @@ public:
   explicit PlayerState(const RawMap &raw_map, const float move_speed = 3.0f, const float rot_speed = 1.5f);
 
   void set_keyboard_state(std::shared_ptr<gp::misc::KeyboardState> keyboard_state);
-  void animate(const std::uint32_t time_elapsed_ms);
+  void animate(const std::uint64_t time_elapsed_ms);
 
   float orientation() const;
   glm::vec2 pos() const;
@@ -33,9 +33,9 @@ private:
   std::shared_ptr<gp::misc::KeyboardState> keyboard_state_{};
 
   float deduce_orientation(const RawMap &raw_map) const;
-  void animate_move_noclip(const std::uint32_t time_elapsed_ms);
-  void animate_move(const std::uint32_t time_elapsed_ms);
-  void animate_rot(const std::uint32_t time_elapsed_ms);
+  void animate_move_noclip(const std::uint64_t time_elapsed_ms);
+  void animate_move(const std::uint64_t time_elapsed_ms);
+  void animate_rot(const std::uint64_t time_elapsed_ms);
 
   int move_direction() const;
   int rot_direction() const;
